Stop using 32-bit time in SaveInfo

+++ src/client/http/GetSaveRequest.cpp
@@ -44,8 +44,8 @@ namespace http
 			}
 			saveInfo = std::make_unique<SaveInfo>(
 				document["ID"].asInt(),
-				document["DateCreated"].asInt(),
-				document["Date"].asInt(),
+				document["DateCreated"].asInt64(),
+				document["Date"].asInt64(),
 				document["ScoreUp"].asInt(),
 				document["ScoreDown"].asInt(),
 				document["ScoreMine"].asInt(),
